mreg

+++ cold-web/src/main/java/com/smartcold/manage/cold/controller/UserController.java
@@ -53,6 +53,7 @@ public class UserController extends BaseController {
 	public Object logout(HttpServletRequest request) {
 		request.getSession().setAttribute("user", null);
 		Cookie[] cookies = request.getCookies();
+		if(cookies==null||cookies.length==0){return true;}
 		for (Cookie cookie : cookies) {
 			if (cookie.getName().equals("token")) {
 				cookieService.deleteCookie(cookie.getValue());
